+++ cli/cli/commands/enclave/upload/upload.go
@@ -27,8 +27,8 @@ const (
 	engineClientCtxKey  = "engine-client"
 )
 
-var EnclaveUploadCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosisCommand{
-	CommandStr:              command_str_consts.EnclaveUploadCmdStr,
+var EnclaveFilesUploadCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosisCommand{
+	CommandStr:              command_str_consts.EnclaveFilesUploadCmdStr,
 	ShortDescription:        "Uploads files to an enclave",
 	LongDescription:         "Uploads the requested files to the enclave so they can be used by modules and services within the enclave",
 	KurtosisBackendContextKey: kurtosisBackendCtxKey,